riceleaf-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
numpy==2.0.1

# PyTorch CPU (เบากว่า GPU และเหมาะกับ Container Apps)
torch==2.4.0
torchvision==0.19.0
